requirements.txt
Flask
dbfread
fpdf2
requests
beautifulsoup4
selenium
webdriver-manager
psycopg2-binary
SQLAlchemy
